Fix bug after restoring VM from saved state
When a VM is restored from saved state, for some (unknown) reason the
order of child windows is different. To work around it, check if the child
window we found is invisible, and use the other child window in that case.

As a workaround, it is also possible to resize the VM window after
restoring so that both scroll bars are visible, and then press Host+A to
auto-resize it back. Afterwards, grabbing will work with an unfixed
version as well.